refactor AllProducts and SingleProduct components to remove container wrapper
