+++ client/components/index.js
@@ -6,5 +6,5 @@
 export { default as Navbar } from './navbar';
 export { default as UserHome } from './user-home';
 export { Login, Signup } from './auth-form';
-export { default as AllProducts } from './AllProductsContainer';
-export { default as SingleProduct } from './SingleProductContainer';
+export { default as AllProducts } from './AllProducts';
+export { default as SingleProduct } from './SingleProduct';
